feat(formatter): extract PremiumEventFormatter into log_fmt/ with Axon-style semantic coloring
Extract the human log formatter from main.rs into crates/lab/src/log_fmt/formatter.rs,
cleaning up main.rs to contain only init_tracing() and the entry point.

Key improvements:
- BTreeMap<String, String> → BTreeMap<&'static str, String>: Field::name() already
  returns &'static str, eliminating unnecessary key allocations per log event.
- Remove vestigial .with_ansi(ansi) call: no-op with a custom FormatEvent impl.
- INFO badge: "1;36" (bold cyan) → "32" (green) to match Axon palette.
- Add semantic style helpers: accent() (color256 111, blue), subject_accent()
  (color256 211, pink), error_value() (red), warn_value() (yellow).
- Apply accent() to tool/prompt/resource_uri/upstream/route fields; error_value()
  to error field; warn_value() to kind at WARN/ERROR level.
- Apply subject_accent() to the subject token (service.action) in each log line.
- Trim priority field list from ~40 to 16 runtime-relevant fields; startup-only
  fields fall to alphabetical catch-all with no visible behavior change.
- Add unit tests for normalize_label, format_field_value, should_skip_field,
  sanitize_field_value (clean, injection, preserved tab/newline).
