+++ crates/lab/src/log_fmt.rs
@@ -1,0 +1,1 @@
+pub(crate) mod formatter;
